creating product details page
